+++ src/json_pointer.erl
@@ -15,7 +15,7 @@
 -module(json_pointer).
 
 -export([parent/1, child/2,
-         parse/1, serialize/1, eval/2]).
+         parse/1, serialize/1, find/2, insert/3, replace/3]).
 
 -export_type([pointer/0, reference_token/0,
               error_reason/0,
@@ -32,6 +32,13 @@
 
 -type evaluation_error_reason() :: invalid_pointer
                                  | {invalid_array_index, reference_token()}.
+
+-type update_fun() ::
+        fun(({root, json:value()} |
+             {object, Parent :: json:value(), json:key()} |
+             {array, Parent :: json:value(), non_neg_integer()} |
+             {array_end, Parent :: json:value()}) ->
+               json:value()).
 
 -spec parent(pointer()) -> pointer().
 parent([]) ->
@@ -95,37 +102,131 @@ serialize_reference_token(<<$~, Token/binary>>, Acc) ->
 serialize_reference_token(<<C, Token/binary>>, Acc) ->
   serialize_reference_token(Token, <<Acc/binary, C>>).
 
--spec eval(binary() | pointer(), json:value())
-                  -> {ok, json:value()} | {error, error_reason()}.
-eval(PointerString, Value) when is_binary(PointerString) ->
+-spec find(binary() | pointer(), json:value())
+           -> {ok, json:value()} | {error, error_reason()}.
+find(PointerString, Value) when is_binary(PointerString) ->
   case parse(PointerString) of
     {ok, Pointer} ->
-      eval(Pointer, Value);
+      find(Pointer, Value);
     {error, Reason} ->
       {error, Reason}
   end;
-eval([], Value) ->
+find([], Value) ->
   {ok, Value};
-eval([Token | Tokens], Value) when is_map(Value) ->
+find([Token | Tokens], Value) when is_map(Value) ->
   case maps:find(Token, Value) of
     {ok, Child} ->
-      eval(Tokens, Child);
+      find(Tokens, Child);
     error ->
       {error, invalid_pointer}
   end;
-eval([<<"-">> | _], Value) when is_list(Value) ->
+find([<<"-">> | _], Value) when is_list(Value) ->
   {error, invalid_pointer};
-eval([Token | Tokens], Value) when is_list(Value) ->
+find([Token | Tokens], Value) when is_list(Value) ->
   try
     binary_to_integer(Token)
   of
     I when I >= 0, I < length(Value) ->
-      eval(Tokens, lists:nth(I+1, Value));
+      find(Tokens, lists:nth(I+1, Value));
     _ ->
       {error, invalid_pointer}
   catch
     error:badarg ->
       {error, {invalid_array_index, Token}}
   end;
-eval(_Pointer, _Value) ->
+find(_Pointer, _Value) ->
   {error, invalid_pointer}.
+
+-spec update(binary() | pointer(), json:value(), update_fun())
+            -> {ok, json:value()} | {error, error_reason()}.
+update(PointerString, Value, F) when is_binary(PointerString) ->
+  case parse(PointerString) of
+    {ok, Pointer} ->
+      update(Pointer, Value, F);
+    {error, Reason} ->
+      {error, Reason}
+  end;
+update(Pointer, Value, F) ->
+  try
+    {ok, do_update(Pointer, Value, F)}
+  catch
+    throw:{error, Reason} ->
+      {error, Reason}
+  end.
+
+-spec do_update(pointer(), json:value(), update_fun()) -> json:value().
+do_update([], Value, F) ->
+  F({root, Value});
+do_update([Token], Value, F) when is_map(Value) ->
+  F({object, Value, Token});
+do_update([Token | Tokens], Value, F) when is_map(Value) ->
+  case maps:find(Token, Value) of
+    {ok, Child} ->
+      Value#{Token => do_update(Tokens, Child, F)};
+    error ->
+      throw({error, invalid_pointer})
+  end;
+do_update([<<"-">> | []], Value, F) when is_list(Value) ->
+  F({array_end, Value});
+do_update([<<"-">> | _], Value, _F) when is_list(Value) ->
+  throw({error, invalid_pointer});
+do_update([Token | Tokens], Value, F) when is_list(Value) ->
+  try
+    binary_to_integer(Token)
+  of
+    I when I >= 0, I < length(Value) ->
+      case Tokens of
+        [] ->
+          F({array, Value, I});
+        _ ->
+          NewValue = do_update(Tokens, lists:nth(I+1, Value), F),
+          {Before, [_ | After]} = lists:split(I, Value),
+          Before ++ [NewValue | After]
+      end;
+    I when I == length(Value) ->
+      case Tokens of
+        [] ->
+          F({array_end, Value});
+        _ ->
+          throw({error, invalid_pointer})
+      end;
+    _ ->
+      throw({error, invalid_pointer})
+  catch
+    error:badarg ->
+      throw({error, {invalid_array_index, Token}})
+  end;
+do_update(_Pointer, _Value, _F) ->
+  throw({error, invalid_pointer}).
+
+-spec insert(pointer(), json:value(), NewValue :: json:value()) ->
+        json:value().
+insert(Pointer, Value, NewValue) ->
+  F = fun
+        ({root, _}) ->
+          NewValue;
+        ({object, Parent, Key}) ->
+          Parent#{Key => NewValue};
+        ({array, Parent, I}) ->
+          {Before, After} = lists:split(I, Parent),
+          Before ++ [NewValue] ++ After;
+        ({array_end, Parent}) ->
+          Parent ++ [NewValue]
+      end,
+  update(Pointer, Value, F).
+
+-spec replace(pointer(), json:value(), NewValue :: json:value()) ->
+        json:value().
+replace(Pointer, Value, NewValue) ->
+  F = fun
+        ({root, _}) ->
+          NewValue;
+        ({object, Parent, Key}) ->
+          Parent#{Key => NewValue};
+        ({array, Parent, I}) ->
+          {Before, [_ | After]} = lists:split(I, Parent),
+          Before ++ [NewValue | After];
+        ({array_end, _}) ->
+          throw({error, invalid_pointer})
+      end,
+  update(Pointer, Value, F).